ai-rag-mircoservice/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astrapy==2.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
deprecation==2.1.0
dnspython==2.7.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-astradb==0.6.0
langchain-community==0.3.21
langchain-core==0.3.63
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.43
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
orjson==3.10.18
packaging==24.2
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uuid6==2024.7.10
yarl==1.20.0
zstandard==0.23.0
